move reusable components into new components directory. fix styles.
